requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
distlib==0.3.5
fastapi==0.79.0
filelock==3.7.1
frozenlist==1.3.1
h11==0.13.0
identify==2.5.3
idna==3.3
motor==3.0.0
multidict==6.0.2
nodeenv==1.7.0
platformdirs==2.5.2
pre-commit==2.20.0
pydantic==1.9.1
PyJWT==2.4.0
pymongo==4.2.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.2
virtualenv==20.16.3
yarl==1.8.1
